Support sub-lists and tables in lists.

+++ LearnDocUtils/DocxToLearn.cs
@@ -42,7 +42,8 @@ namespace LearnDocUtils
             {
                 if (!debug)
                 {
-                    File.Delete(markdownFile);
+                    if (File.Exists(markdownFile))
+                        File.Delete(markdownFile);
                 }
             }
         }
